requirements.txt
py-cpuinfo==9.0.0
python-dotenv==1.0.0
scikit-learn==1.3.2
openpyxl==3.1.2
datetime==5.4
urllib3==1.24.3
openai==1.7.0
langchain==0.1.0
einops==0.7.0
scipy==1.11.4
deepspeed==0.3.16
openai-whisper==20230314
torch
ffmpeg-python~=0.2.0
hanziconv==0.3.2
numpy==1.26.2
noisereduce==3.0.0
bilibili-api-python==16.2.0
ass==0.5.4
pysubs2==1.6.1
librosa==0.10.1
chardet==5.1.0
pysrt==1.1.2
pandas==2.1.4
audioread==3.0.1
onnxruntime==1.16.3
tqdm>=4.63.1
soundfile>=0.12.1

transformers>=4.36.2
datasets>=2.14.3
accelerate>=0.21.0
peft>=0.7.0
trl>=0.7.6
gradio>=3.38.0,<4.0.0
sentencepiece
protobuf
tiktoken
jieba
rouge-chinese
nltk
uvicorn
pydantic
fastapi
sse-starlette
matplotlib
faiss-cpu
FlagEmbedding
httpx[socks]
